14/02/2021 Subida desde el pc fijo. Hacer que de la ventana de nuevo mapa, después de guardar el mapa en la BD, vaya a la ventana de Editar Mapa. Hay que comprobar que aparecen todos los datos. Y hay que mostrar el ID del mapa

+++ app/src/main/java/com/pgsanchez/whereismymap/repository/MapRepository.java
@@ -10,10 +10,10 @@ public interface MapRepository {
     // Métodos:
 
     List<Map> getAllMaps();
-    Map getMapById(int id);
+    Map getMapById(long id);
     List<Map> getMapsByName(String name);
 
-    int insertMap(Map mapa);
+    long insertMap(Map mapa);
     void updateMap(Map mapa);
     void deleteMap(Map mapa);
     void deleteMap(int id);
